feat: #338 댓글 좋아요 기능 추가

+++ android/TodokTodok/domain/src/main/java/com/team/domain/model/Reply.kt
@@ -8,6 +8,6 @@ data class Reply(
     val createdAt: LocalDateTime,
     val likeCount: Int,
     val user: User,
-    val replyId: Int,
+    val replyId: Long,
     val isLikedByMe: Boolean,
 )
